Add support for systemd socket activation
The smtproxy can now be run from a .socket unit file by systemd,
meaning it's spawned on demand when the first mail client connects.
This also means smtpproxy does not require root privileges anymore,
as it does not bind to port 25.

Careful, the user it runs as still needs to be able to read any
certificates you specify.

Fixes #3